The part of PBFTEngine was modified
I corrected some grammatical errors I thought.
And for the sequence of processe, I think there's a mistake in original passage.
But I still have a problem that is "transactions will be deleted from txPool". I don't know if it's right to put it last of the whole process.
Looking forward to your reply. THX :-)